Changed the code structure and pr description (#89)
changed the code structure and pr description
