+++ .github/workflows/auto-submit-pr.yaml
@@ -1,17 +1,17 @@
 on:
   push:
     branches:
-      - action-dev
+      - main
   schedule:
-    - cron: '10 * * * *'
+    - cron: '0 2 * * *'
 
-name: auto collect info and create pull request
+name: auto-collector
 jobs:
   auto-createPR:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
-      - name: auto collect periodically  
+      - uses: actions/checkout@v3
+      - name: Launch auto-collector periodically  
         run: make collect
       - name: Run read-yaml action
         id: yaml-data
@@ -24,12 +24,12 @@ jobs:
           echo "${{ steps.yaml-data.outputs.updated }}"
           
       - name: Create Pull Request
-        if: ${{steps.yaml-data.outputs.updated == 'true' }}
+        if: ${{ steps.yaml-data.outputs.updated == 'true' }}
         uses: peter-evans/create-pull-request@v3
         with:
-          commit-message: update config file
-          title: The edgeX has been updated
+          commit-message: discover edgex new version
+          title: "[auto-collector] Discover edgex new version"
           body: >
             This PR is auto-generated by github action,please check and merge the config file in time.
-          labels: update file, automated pr
-          branch: auto-generated-branch
+          labels: auto-collector, automated pr
+          branch: auto-collector-generated-branch
