+++ tools/collector/edgex/error.go
